Moved emiting queries methods to MsgVariant
