+++ sylvia-derive/src/querier.rs
@@ -1,5 +1,4 @@
-use convert_case::{Case, Casing};
-use proc_macro2::{Ident, TokenStream};
+use proc_macro2::TokenStream;
 use proc_macro_error::emit_error;
 use quote::quote;
 use syn::parse::{Parse, Parser};
@@ -8,7 +7,7 @@ use syn::{GenericParam, ImplItem, ItemImpl, ItemTrait, TraitItem};
 
 use crate::check_generics::CheckGenerics;
 use crate::crate_module;
-use crate::message::{MsgField, MsgVariant};
+use crate::message::MsgVariant;
 use crate::parser::{MsgAttr, MsgType};
 
 pub struct Querier<'a> {
@@ -76,29 +75,8 @@ impl<'a> Querier<'a> {
         let sylvia = crate_module();
         let Self { variants } = self;
 
-        let methods_impl = variants.iter().map(|variant| {
-            let MsgVariant { name, fields, return_type, .. } = variant;
-
-            let parameters = fields.iter().map(MsgField::emit_without_attrs);
-            let fields_names = fields.iter().map(MsgField::name);
-            let variant_name = Ident::new(&name.to_string().to_case(Case::Snake), name.span());
-
-            quote! {
-                fn #variant_name(&self, #(#parameters),*) -> Result< #return_type, #sylvia:: cw_std::StdError> {
-                    let query = QueryMsg:: #variant_name (#(#fields_names),*);
-                    self.querier.query_wasm_smart(self.contract, &query)
-                }
-            }
-        });
-
-        let methods_declaration = variants.iter().map(|variant| {
-            let MsgVariant { name, fields, return_type, .. } = variant;
-
-            let parameters = fields.iter().map(MsgField::emit_without_attrs);
-            let variant_name = Ident::new(&name.to_string().to_case(Case::Snake), name.span());
-
-            quote! {
-                fn #variant_name(&self, #(#parameters),*) -> Result< #return_type, #sylvia:: cw_std::StdError>;}});
+        let methods_impl = variants.iter().map(MsgVariant::emit_querier_impl);
+        let methods_declaration = variants.iter().map(MsgVariant::emit_querier_declaration);
 
         quote! {
             pub struct BoundQuerier<'a, C: #sylvia ::cw_std::CustomQuery> {
